+++ packages/brick_offline_first_build/lib/src/offline_first_sqlite_generators.dart
@@ -50,6 +50,15 @@ class OfflineFirstSqliteSerialize extends SqliteSerialize {
       fieldAnnotation: fieldAnnotation,
     );
   }
+
+  @override
+  String uniqueValueForField(fieldName, {required checker}) {
+    if ((checker as OfflineFirstChecker).hasSerdes) {
+      return '$fieldName.toSqlite()';
+    }
+
+    return super.uniqueValueForField(fieldName, checker: checker);
+  }
 }
 
 class OfflineFirstSqliteDeserialize extends SqliteDeserialize {