+++ packages/brick_offline_first/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
 
+## 3.2.0
+
 * Apply desired policy to all association fetches during `OfflineFirstRepository#get` requests. Fixes #371.
+* Upgrade minimum Dart to 2.18
 
 ## 3.1.0