+++ packages/brick_offline_first_with_graphql_build/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 3.2.0
+
 * Apply standardized lints
 * Upgrade minimum Dart to 2.18