+++ packages/brick_sqlite_generators/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Unreleased
 
+## 3.1.0
+
 * Apply standardized lints
 * Upgrade minimum Dart to 2.18
+* Add `SqliteSerialize#uniqueValueForField` for advanced control of unique field genertion
 
 ## 3.0.2